Refactor static Strategies to inherit from a common base class
