+++ blackjack/strategies/base_static_strategy.py
@@ -8,36 +8,23 @@ from blackjack.strategies.base_strategy import BaseStrategy
 DIRECTORY = os.path.dirname(os.path.realpath(__file__))
 
 
-class StaticStrategy(BaseStrategy):
-    """Default predetermined Strategy for optimal odds."""
+class BaseStaticStrategy(BaseStrategy):
+    """
+    Base predetermined Strategy from which other static Strategies can be derrived.
+    Note that concrete static Strategies must implement the required BaseStrategy methods omitted here.
+    """
 
-    def __init__(self,
-                 split_csv=f"{DIRECTORY}/default_split.csv",
-                 soft_csv=f"{DIRECTORY}/default_soft.csv",
-                 hard_csv=f"{DIRECTORY}/default_hard.csv"
-                 ):
+    def __init__(self, strategy_name):
         super().__init__()
-        self.split_df = self._load_df(split_csv)
-        self.soft_df = self._load_df(soft_csv)
-        self.hard_df = self._load_df(hard_csv)
-
-    # --- Helper Methods --- #
+        self.split_df = self._load_df(strategy_name, 'split')
+        self.soft_df = self._load_df(strategy_name, 'soft')
+        self.hard_df = self._load_df(strategy_name, 'hard')
 
     @staticmethod
-    def _load_df(csv_path):
+    def _load_df(strategy_name, csv_type):
         """Load a DataFrame from a CSV for determining actions."""
+        csv_path = f"{DIRECTORY}/csv/{strategy_name}/{csv_type}.csv"
         return read_csv(csv_path, index_col=0)
-
-    # --- Required Methods --- #
-
-    def wants_to_change_wager(self):
-        """Get a yes/no response (bool) for whether the gambler wants to change their auto-wager."""
-        return False
-
-    def get_new_auto_wager(self):
-        """Get a new auto-wager amount (float)."""
-        # Will not change the auto-wager in this strategy
-        pass
 
     def get_hand_action(self, hand, options, dealer_upcard):
         """Get the action to take on the hand ('Hit', 'Stand', etc.)"""
@@ -62,11 +49,3 @@ class StaticStrategy(BaseStrategy):
             return 'Hit'
         
         return action
-
-    def wants_even_money(self):
-        """Get a yes/no response (bool) for whether a the gambler wants to take even money for a blackjack when facing an Ace."""
-        return False
-
-    def wants_insurance(self):
-        """Get a yes/no response (bool) for whether a user wants to make an insurance bet when facing an Ace."""
-        return False
